Remove a few interfaces
