+++ api/metric/sdkhelpers.go
@@ -20,10 +20,8 @@ import (
 	"go.opentelemetry.io/otel/api/core"
 )
 
-// MeterImpl is a convenient interface for SDK and test
-// implementations that would provide a `Meter` but do not wish to
-// re-implement the API's type-safe interfaces.  Helpers provided in
-// this package will construct a `Meter` given a `MeterImpl`.
+// MeterImpl is the interface an SDK must implement to supply a Meter
+// implementation.
 type MeterImpl interface {
 	// RecordBatch atomically records a batch of measurements.
 	RecordBatch(context.Context, []core.KeyValue, ...Measurement)
@@ -83,33 +81,19 @@ type BoundSyncImpl interface {
 // asynchronous instrument (e.g., Observer instruments).
 type AsyncImpl interface {
 	InstrumentImpl
-
-	// Note: An `Unregister()` API could be supported here.
 }
 
-// Meter implements the `Meter` interface given a
-// `MeterImpl` implementation.
-type Meter struct {
-	impl        MeterImpl
-	libraryName string
-}
-
-// int64ObserverResult is an adapter for int64-valued asynchronous
-// callbacks.
-type int64ObserverResult struct {
+// Int64ObserverResult is passed to an observer callback to capture
+// observations for one asynchronous integer metric instrument.
+type Int64ObserverResult struct {
 	observe func(core.Number, []core.KeyValue)
 }
 
-// float64ObserverResult is an adapter for float64-valued asynchronous
-// callbacks.
-type float64ObserverResult struct {
+// Float64ObserverResult is passed to an observer callback to capture
+// observations for one asynchronous floating point metric instrument.
+type Float64ObserverResult struct {
 	observe func(core.Number, []core.KeyValue)
 }
-
-var (
-	_ Int64ObserverResult   = int64ObserverResult{}
-	_ Float64ObserverResult = float64ObserverResult{}
-)
 
 // Configure is a helper that applies all the options to a Config.
 func Configure(opts []Option) Config {
@@ -129,17 +113,12 @@ func WrapMeterImpl(impl MeterImpl, libraryName string) Meter {
 	}
 }
 
+// MeterImpl returns the underlying MeterImpl of this Meter.
 func (m Meter) MeterImpl() MeterImpl {
 	return m.impl
 }
 
-func (m Meter) RecordBatch(ctx context.Context, ls []core.KeyValue, ms ...Measurement) {
-	if m.impl == nil {
-		return
-	}
-	m.impl.RecordBatch(ctx, ls, ms...)
-}
-
+// newSync constructs one new synchronous instrument.
 func (m Meter) newSync(name string, metricKind Kind, numberKind core.NumberKind, opts []Option) (SyncImpl, error) {
 	if m.impl == nil {
 		return NoopSync{}, nil
@@ -147,11 +126,6 @@ func (m Meter) newSync(name string, metricKind Kind, numberKind core.NumberKind,
 	desc := NewDescriptor(name, metricKind, numberKind, opts...)
 	desc.config.LibraryName = m.libraryName
 	return m.impl.NewSyncInstrument(desc)
-}
-
-func (m Meter) NewInt64Counter(name string, opts ...Option) (Int64Counter, error) {
-	return wrapInt64CounterInstrument(
-		m.newSync(name, CounterKind, core.Int64NumberKind, opts))
 }
 
 // wrapInt64CounterInstrument returns an `Int64Counter` from a
@@ -163,11 +137,6 @@ func wrapInt64CounterInstrument(syncInst SyncImpl, err error) (Int64Counter, err
 	return Int64Counter{syncInstrument: common}, err
 }
 
-func (m Meter) NewFloat64Counter(name string, opts ...Option) (Float64Counter, error) {
-	return wrapFloat64CounterInstrument(
-		m.newSync(name, CounterKind, core.Float64NumberKind, opts))
-}
-
 // wrapFloat64CounterInstrument returns an `Float64Counter` from a
 // `SyncImpl`.  An error will be generated if the
 // `SyncImpl` is nil (in which case a No-op is substituted),
@@ -175,11 +144,6 @@ func (m Meter) NewFloat64Counter(name string, opts ...Option) (Float64Counter, e
 func wrapFloat64CounterInstrument(syncInst SyncImpl, err error) (Float64Counter, error) {
 	common, err := checkNewSync(syncInst, err)
 	return Float64Counter{syncInstrument: common}, err
-}
-
-func (m Meter) NewInt64Measure(name string, opts ...Option) (Int64Measure, error) {
-	return wrapInt64MeasureInstrument(
-		m.newSync(name, MeasureKind, core.Int64NumberKind, opts))
 }
 
 // wrapInt64MeasureInstrument returns an `Int64Measure` from a
@@ -191,11 +155,6 @@ func wrapInt64MeasureInstrument(syncInst SyncImpl, err error) (Int64Measure, err
 	return Int64Measure{syncInstrument: common}, err
 }
 
-func (m Meter) NewFloat64Measure(name string, opts ...Option) (Float64Measure, error) {
-	return wrapFloat64MeasureInstrument(
-		m.newSync(name, MeasureKind, core.Float64NumberKind, opts))
-}
-
 // wrapFloat64MeasureInstrument returns an `Float64Measure` from a
 // `SyncImpl`.  An error will be generated if the
 // `SyncImpl` is nil (in which case a No-op is substituted),
@@ -205,6 +164,7 @@ func wrapFloat64MeasureInstrument(syncInst SyncImpl, err error) (Float64Measure,
 	return Float64Measure{syncInstrument: common}, err
 }
 
+// newAsync constructs one new asynchronous instrument.
 func (m Meter) newAsync(name string, mkind Kind, nkind core.NumberKind, opts []Option, callback func(func(core.Number, []core.KeyValue))) (AsyncImpl, error) {
 	if m.impl == nil {
 		return NoopAsync{}, nil
@@ -212,20 +172,6 @@ func (m Meter) newAsync(name string, mkind Kind, nkind core.NumberKind, opts []O
 	desc := NewDescriptor(name, mkind, nkind, opts...)
 	desc.config.LibraryName = m.libraryName
 	return m.impl.NewAsyncInstrument(desc, callback)
-}
-
-func (m Meter) RegisterInt64Observer(name string, callback Int64ObserverCallback, opts ...Option) (Int64Observer, error) {
-	if callback == nil {
-		return wrapInt64ObserverInstrument(NoopAsync{}, nil)
-	}
-	return wrapInt64ObserverInstrument(
-		m.newAsync(name, ObserverKind, core.Int64NumberKind, opts,
-			func(observe func(core.Number, []core.KeyValue)) {
-				// Note: this memory allocation could be avoided by
-				// using a pointer to this object and mutating it
-				// on each collection interval.
-				callback(int64ObserverResult{observe})
-			}))
 }
 
 // wrapInt64ObserverInstrument returns an `Int64Observer` from a
@@ -237,17 +183,6 @@ func wrapInt64ObserverInstrument(asyncInst AsyncImpl, err error) (Int64Observer,
 	return Int64Observer{asyncInstrument: common}, err
 }
 
-func (m Meter) RegisterFloat64Observer(name string, callback Float64ObserverCallback, opts ...Option) (Float64Observer, error) {
-	if callback == nil {
-		return wrapFloat64ObserverInstrument(NoopAsync{}, nil)
-	}
-	return wrapFloat64ObserverInstrument(
-		m.newAsync(name, ObserverKind, core.Float64NumberKind, opts,
-			func(observe func(core.Number, []core.KeyValue)) {
-				callback(float64ObserverResult{observe})
-			}))
-}
-
 // wrapFloat64ObserverInstrument returns an `Float64Observer` from a
 // `AsyncImpl`.  An error will be generated if the
 // `AsyncImpl` is nil (in which case a No-op is substituted),
@@ -255,12 +190,4 @@ func (m Meter) RegisterFloat64Observer(name string, callback Float64ObserverCall
 func wrapFloat64ObserverInstrument(asyncInst AsyncImpl, err error) (Float64Observer, error) {
 	common, err := checkNewAsync(asyncInst, err)
 	return Float64Observer{asyncInstrument: common}, err
-}
-
-func (io int64ObserverResult) Observe(value int64, labels ...core.KeyValue) {
-	io.observe(core.NewInt64Number(value), labels)
-}
-
-func (fo float64ObserverResult) Observe(value float64, labels ...core.KeyValue) {
-	fo.observe(core.NewFloat64Number(value), labels)
 }
